requirements.txt
pytest
pytest-cov
apache-airflow[http,cncf.kubernetes] >=2.0.0
SQLAlchemy==1.3.23
Flask-SQLAlchemy==2.4.4
markupsafe>=1.1.1,<2.1.0